desktop/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# File Handling
python-multipart>=0.0.6
aiofiles>=23.2.1

# Database
sqlalchemy>=2.0.0

# OpenAI API
openai>=1.3.0

# HTTP Client (for async requests)
httpx>=0.25.0

# Google Drive API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0

# Encryption (for API key transfer)
cryptography>=41.0.0
